+++ Node.java
@@ -14,7 +14,7 @@ public class Node {
     public boolean independentVar = false;
 
     //constant int member variable
-    public int constant;
+    public float constant;
 
     //constructor for a root Node
     public Node()
